+++ x10/configuration.py
@@ -1,3 +1,4 @@
+# FIXME: Move one level up
 from dataclasses import dataclass
