+++ x10/core/stark_account.py
@@ -1,19 +1,12 @@
-from decimal import Decimal
-from typing import Dict, List, Optional, Tuple
+from typing import Dict, Tuple
 
 from fast_stark_crypto import sign
-from pydantic import AliasChoices, Field
 
-from x10.perpetual.balances import BalanceModel
 from x10.perpetual.fees import TradingFeeModel
-from x10.perpetual.orders import OpenOrderModel
-from x10.perpetual.positions import PositionModel
-from x10.perpetual.trades import AccountTradeModel
-from x10.utils.model import X10BaseModel
 from x10.utils.string import is_hex_string
 
 
-class StarkPerpetualAccount:
+class StarkAccount:
     __vault: int
     __private_key: int
     __public_key: int
@@ -54,33 +47,3 @@ class StarkPerpetualAccount:
 
     def sign(self, msg_hash: int) -> Tuple[int, int]:
         return sign(private_key=self.__private_key, msg_hash=msg_hash)
-
-
-class AccountStreamDataModel(X10BaseModel):
-    orders: Optional[List[OpenOrderModel]] = None
-    positions: Optional[List[PositionModel]] = None
-    trades: Optional[List[AccountTradeModel]] = None
-    balance: Optional[BalanceModel] = None
-
-
-class AccountLeverage(X10BaseModel):
-    market: str
-    leverage: Decimal
-
-
-class AccountModel(X10BaseModel):
-    id: int = Field(validation_alias=AliasChoices("accountId", "id"), serialization_alias="id")
-    description: str
-    account_index: int
-    status: str
-    l2_key: str
-    l2_vault: int
-    bridge_starknet_address: Optional[str] = None
-
-
-class ApiKeyResponseModel(X10BaseModel):
-    key: str
-
-
-class ApiKeyRequestModel(X10BaseModel):
-    description: str
