+++ x10/models/balance.py
@@ -1,6 +1,6 @@
 from decimal import Decimal
 
-from x10.utils.model import X10BaseModel
+from x10.models.base import X10BaseModel
 
 
 class BalanceModel(X10BaseModel):
